# Conflicts:
#	parallel-based-sorting-algorithms/src/test/java/michaeljosh/performanceTest.java
#	parallel-based-sorting-algorithms/target/classes/michaeljosh/parallelBrickSort.class

+++ parallel-based-sorting-algorithms/src/test/java/michaeljosh/radixSortTest.java
@@ -9,8 +9,8 @@ import michaeljosh.RadixSort.SequentialRadixSort;
 public class radixSortTest {
     @Test
     public void testRadixSort() {
-        int[] unsortedArray = { 4, 5, 6, 7, 3, 1, 8, 2, 9, 10 };
-        int[] sortedArray = { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
+        long[] unsortedArray = { 4, 5, 6, 7, 3, 1, 8, 2, 9, 10 };
+        long[] sortedArray = { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
 
         SequentialRadixSort.radixSort(unsortedArray); // Assuming you have a Mergesort class with a static sort method
